Split test_pyopnsense
