+++ tests/pyopnsense/__init__.py
@@ -1,0 +1,1 @@
+"""pytest for pyopnsense."""
