+++ tests/pyopnsense/test_exceptions.py
@@ -1,0 +1,11 @@
+"""Tests for `pyopnsense.exceptions`."""
+
+import pytest
+
+from custom_components.opnsense import pyopnsense
+
+
+def test_voucher_server_error() -> None:
+    """Raise VoucherServerError to ensure the exception class exists."""
+    with pytest.raises(pyopnsense.VoucherServerError):
+        raise pyopnsense.VoucherServerError
